requirements.txt
torch==2.2.2
torchvision==0.17.2
opencv-python==4.8.1.78
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
kafka-python==2.0.2
prometheus-client==0.17.1
numpy==1.26.1
onnx==1.15.0
onnxruntime==1.17.1 